Create README.md
This is a FUll Stack Music Application that has the different features that a online music app posses, the features includes Registering an Account, Logging into Account, Songs Lists,Play Lists, Adding a New Song, Creation of Playlist, Subscription option, Payment Gateway Integration. And the technology stack used to build is Java and springboot as Backend, MySQL as Database and HTML,CSS and JavaScript for Frontend, Razorpay API for payment gateway, GitHub for storing the songs.